Move fixPosition into CutoutViewIndicator proper
It was pretty specialised to begin with, and depends on a tentative field, so I guess it makes sense to not tie the method to ViewPager any more than necessary.
Might become helpful for other Generators or StateProxies in the future (though I can only think of one such case right now). We'll see.